+++ convex/_generated/api.d.ts
@@ -15,12 +15,14 @@ import type * as aggregateWorkflows from "../aggregateWorkflows.js";
 import type * as aggregates from "../aggregates.js";
 import type * as bookmark from "../bookmark.js";
 import type * as customQuizzes from "../customQuizzes.js";
+import type * as customQuizzesCreation from "../customQuizzesCreation.js";
 import type * as groups from "../groups.js";
 import type * as http from "../http.js";
 import type * as init from "../init.js";
 import type * as migrations from "../migrations.js";
 import type * as presetQuizzes from "../presetQuizzes.js";
 import type * as questions from "../questions.js";
+import type * as questionsAggregateSync from "../questionsAggregateSync.js";
 import type * as quiz from "../quiz.js";
 import type * as quizSessions from "../quizSessions.js";
 import type * as subthemes from "../subthemes.js";
@@ -53,12 +55,14 @@ declare const fullApi: ApiFromModules<{
   aggregates: typeof aggregates;
   bookmark: typeof bookmark;
   customQuizzes: typeof customQuizzes;
+  customQuizzesCreation: typeof customQuizzesCreation;
   groups: typeof groups;
   http: typeof http;
   init: typeof init;
   migrations: typeof migrations;
   presetQuizzes: typeof presetQuizzes;
   questions: typeof questions;
+  questionsAggregateSync: typeof questionsAggregateSync;
   quiz: typeof quiz;
   quizSessions: typeof quizSessions;
   subthemes: typeof subthemes;
